+++ src/components/index.js
@@ -2,5 +2,14 @@ import Footer from "./Footer";
 import Header from "./Header";
 import ProtectedRoutes from "./ProtectedRoutes";
 import PlacesAutocomplete from "./PlacesAutocomplete";
+import InstantMessage from "./InstantMessage";
+import BookRideComponent from "./BookRideComponent";
 
-export { Footer, Header, ProtectedRoutes, PlacesAutocomplete };
+export {
+  Footer,
+  Header,
+  ProtectedRoutes,
+  PlacesAutocomplete,
+  InstantMessage,
+  BookRideComponent,
+};
